lambda/post-confirmation/requirements.txt
# Lambda runtime includes boto3, no need to bundle
python-ulid==3.1.0
typing_extensions>=4.0.0
